return the task ip obtained from the cni plugin execution while creating task bridge
