+++ agent/ecscni/plugin.go
@@ -21,6 +21,8 @@ import (
 	"path/filepath"
 	"time"
 
+	cnitypes "github.com/containernetworking/cni/pkg/types"
+
 	"github.com/aws/amazon-ecs-agent/agent/logger"
 	"github.com/cihub/seelog"
 	"github.com/containernetworking/cni/libcni"
@@ -88,6 +90,58 @@ func (client *cniClient) SetupNS(
 	ctx, cancel := context.WithTimeout(ctx, timeout)
 	defer cancel()
 	return client.setupNS(ctx, cfg)
+}
+
+// setupNS is the called by SetupNS to setup the task namespace by invoking ADD for given CNI configurations
+func (client *cniClient) setupNS(ctx context.Context, cfg *Config) (*current.Result, error) {
+	seelog.Debugf("[ECSCNI] Setting up the container namespace %s", cfg.ContainerID)
+
+	var bridgeResult cnitypes.Result
+	runtimeConfig := libcni.RuntimeConf{
+		ContainerID: cfg.ContainerID,
+		NetNS:       cfg.ContainerNetNS,
+	}
+
+	// Execute all CNI network configurations serially, in the given order.
+	for _, networkConfig := range cfg.NetworkConfigs {
+		cniNetworkConfig := networkConfig.CNINetworkConfig
+		seelog.Debugf("[ECSCNI] Adding network %s type %s in the container namespace %s",
+			cniNetworkConfig.Network.Name,
+			cniNetworkConfig.Network.Type,
+			cfg.ContainerID)
+		runtimeConfig.IfName = networkConfig.IfName
+		result, err := client.libcni.AddNetwork(ctx, cniNetworkConfig, &runtimeConfig)
+		if err != nil {
+			return nil, errors.Wrap(err, "add network failed")
+		}
+		// Save the result object from the bridge plugin execution. We need this later
+		// for inferring what IPv4 address was used to bring up the veth pair for task.
+		if isBridgePluginExecution(cniNetworkConfig) {
+			bridgeResult = result
+		}
+
+		seelog.Debugf("[ECSCNI] Completed adding network %s type %s in the container namespace %s",
+			cniNetworkConfig.Network.Name,
+			cniNetworkConfig.Network.Type,
+			cfg.ContainerID)
+	}
+
+	seelog.Debugf("[ECSCNI] Completed setting up the container namespace: %s", bridgeResult.String())
+
+	if _, err := bridgeResult.GetAsVersion(currentCNISpec); err != nil {
+		seelog.Warnf("[ECSCNI] Unable to convert result to spec version %s; error: %v; result is of version: %s",
+			currentCNISpec, err, bridgeResult.Version())
+		return nil, err
+	}
+	var curResult *current.Result
+	curResult, ok := bridgeResult.(*current.Result)
+	if !ok {
+		return nil, errors.Errorf(
+			"cni setup: unable to convert result to expected version '%s'",
+			bridgeResult.String())
+	}
+
+	return curResult, nil
 }
 
 // CleanupNS will clean up the container namespace, including remove the veth
